# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Preferences": {
        "properties": {
          "precipitation_unit": {
            "enum": [
              "in",
              "mm"
            ],
            "title": "Precipitation Unit",
            "type": "string"
          },
          "show_advanced_forecast": {
            "title": "Show Advanced Forecast",
            "type": "boolean"
          },
          "temperature_unit": {
            "enum": [
              "F",
              "C"
            ],
            "title": "Temperature Unit",
            "type": "string"
          },
          "wind_unit": {
            "enum": [
              "mph",
              "kmh",
              "mps",
              "kt"
            ],
            "title": "Wind Unit",
            "type": "string"
          }
        },
        "required": [
          "temperature_unit",
          "precipitation_unit",
          "wind_unit",
          "show_advanced_forecast"
        ],
        "title": "Preferences",
        "type": "object"
      },
      "PreferencesUpdate": {
        "properties": {
          "precipitation_unit": {
            "anyOf": [
              {
                "enum": [
                  "in",
                  "mm"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation Unit"
          },
          "show_advanced_forecast": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Advanced Forecast"
          },
          "temperature_unit": {
            "anyOf": [
              {
                "enum": [
                  "F",
                  "C"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Unit"
          },
          "wind_unit": {
            "anyOf": [
              {
                "enum": [
                  "mph",
                  "kmh",
                  "mps",
                  "kt"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Unit"
          }
        },
        "title": "PreferencesUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WeatherNext Personal Weather API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/preferences": {
      "patch": {
        "operationId": "update_preferences_api_v1_me_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Preferences",
        "tags": [
          "visitor"
        ]
      }
    },
...
  }
}